Split runtime and compiler Rust modules
